pyproject.toml
[tool.poetry]
name = "Discovery"
version = "0.1.0"
description = "Connect app to server"
authors = ["Cooperative Code Commun"]
license = "AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-browser-reload = "^1.11"
Markdown = "^3.4.4"
cryptography = "^41.0.4"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
django-extensions = "^3.2.3"
ipdb = "^0.13.13"
ipython = "^8.11.0"
werkzeug = "^3.0.1"
gunicorn = "^21.2.0"
djangorestframework-api-key = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
